+++ .babelrc.js
@@ -11,7 +11,6 @@
 
 const Backtrack = require('@backtrack/core');
 const nodeVersion = require('@backtrack/preset-node/lib/utils/node-version');
-const packageId = require('@backtrack/preset-node/lib/utils/package-id');
 
 const { pkg, configManager } = new Backtrack();
 
@@ -32,14 +31,6 @@ const babel = {
         pkg.resolve(backtrackId, 'babel-plugin-dynamic-import-node'),
         pkg.resolve(backtrackId, '@babel/plugin-proposal-class-properties'),
         pkg.resolve(backtrackId, '@babel/plugin-transform-strict-mode'),
-    ],
-    overrides: [
-        {
-            test: [`./src/${packageId}.js`, `./src/${packageId}.ts`],
-            plugins: [
-                pkg.resolve(backtrackId, 'babel-plugin-add-module-exports'),
-            ],
-        },
     ],
 };